requirements.txt
kafka-python @ git+https://github.com/dpkp/kafka-python.git
